+++ stories/templates/validator/component.template.js
@@ -1,0 +1,5 @@
+export const template = `
+<div class="validator-base">
+  <st-input v-model="inputValue" :validator="inputValidator" />
+</div>
+`;